displays multicodec in links listing
