+++ src/formatters/index.js
@@ -1,0 +1,2 @@
+exports.dag = require('./dag')
+exports.links = require('./links')
